Move timer to dedicated file.
