+++ sys/aarch64/rootdev.c
@@ -7,93 +7,7 @@
 #include <aarch64/interrupt.h>
 #include <sys/kmem.h>
 #include <sys/pmap.h>
-#include <sys/timer.h>
 #include <sys/interrupt.h>
-
-/*
- * TODO(pj): This is only temporary solution. Timer should be initialized by
- * device subsystem.
- */
-
-#define CNTCTL_ENABLE 1
-
-static int timer_start(timer_t *tm, unsigned flags, const bintime_t start,
-                       const bintime_t period) {
-  uint64_t step = bintime_mul(period, tm->tm_frequency).sec;
-
-  WITH_INTR_DISABLED {
-    uint64_t count = READ_SPECIALREG(cntpct_el0);
-    WRITE_SPECIALREG(cntp_cval_el0, count + step);
-    WRITE_SPECIALREG(cntp_ctl_el0, CNTCTL_ENABLE);
-
-    /*
-     * TODO(pj): This should be defined in one header for machine dependent
-     * AArch64 code.
-     */
-    __asm__ volatile("isb" ::: "memory");
-  }
-
-  return 0;
-}
-
-static int timer_stop(timer_t *tm) {
-  WRITE_SPECIALREG(cntp_ctl_el0, 0);
-  return 0;
-}
-
-static bintime_t timer_gettime(timer_t *tm) {
-  uint64_t count = READ_SPECIALREG(cntpct_el0);
-  bintime_t val = bintime_mul(tm->tm_min_period, count);
-  return val;
-}
-
-static timer_t timer = (timer_t){
-  .tm_name = "arm-cpu-timer",
-  .tm_flags = TMF_PERIODIC,
-  .tm_start = timer_start,
-  .tm_stop = timer_stop,
-  .tm_gettime = timer_gettime,
-  .tm_priv = NULL,
-};
-
-static vaddr_t va;
-
-void intr_tick(void) {
-  tm_trigger(&timer);
-
-  /*
-   * https://developer.arm.com/docs/ddi0595/h/aarch64-system-registers/cntp_cval_el0
-   */
-  uint64_t count = READ_SPECIALREG(cntpct_el0);
-  uint64_t freq = READ_SPECIALREG(cntfrq_el0);
-  WRITE_SPECIALREG(cntp_cval_el0, count + freq);
-
-  __asm__ volatile("isb" ::: "memory");
-}
-
-static void _init_clock(void) {
-  /* TODO(pj): This memory should be managed by bus_space_map. */
-  va = kva_alloc(PAGESIZE);
-  assert((void *)va != NULL);
-
-  paddr_t pa = BCM2836_ARM_LOCAL_BASE;
-  pmap_kenter(va, pa, VM_PROT_READ | VM_PROT_WRITE, PMAP_NOCACHE);
-
-  size_t offset = BCM2836_LOCAL_TIMER_IRQ_CONTROLN(0);
-  volatile uint32_t *timerp = (uint32_t *)(va + offset);
-  *timerp = *timerp | (1 << BCM2836_INT_CNTPNSIRQ);
-}
-
-void init_arm_timer(void) {
-  uint64_t freq = READ_SPECIALREG(cntfrq_el0);
-
-  timer.tm_frequency = freq;
-  timer.tm_min_period = HZ2BT(freq);
-  /* TODO(pj): Document what this value means. */
-  timer.tm_max_period = bintime_mul(HZ2BT(freq), 1LL << 30);
-  tm_register(&timer);
-  tm_select(&timer);
-}
 
 typedef struct rootdev {
 } rootdev_t;
@@ -139,6 +53,8 @@ static void rootdev_intr_teardown(device_t *dev, intr_handler_t *handler) {
   aarch64_intr_teardown(handler);
 }
 
+extern int arm_timer_attach(device_t *dev);
+
 static int rootdev_attach(device_t *bus) {
   rman_init(&rm_mem[0], "BCM2835 peripherials", BCM2835_PERIPHERALS_BASE,
             BCM2835_PERIPHERALS_BASE + BCM2835_PERIPHERALS_SIZE - 1, RT_MEMORY);
@@ -149,6 +65,11 @@ static int rootdev_attach(device_t *bus) {
   rootdev_local_handle = kva_alloc(BCM2836_ARM_LOCAL_SIZE);
   pmap_kenter(rootdev_local_handle, BCM2836_ARM_LOCAL_BASE,
               VM_PROT_READ | VM_PROT_WRITE, PMAP_NOCACHE);
+
+  /* TODO(pj) do it in correct way. */
+  device_t *timer = device_add_child(bus, bus->devclass, -1);
+  arm_timer_attach(timer);
+
   return 0;
 }
 
@@ -202,6 +123,4 @@ DEVCLASS_ENTRY(root, rootdev);
 
 void init_devices(void) {
   device_attach(&rootdev);
-  /* TODO(pj) move it to other place. */
-  _init_clock();
 }
